Добавлены сервисы для объявлений

+++ src/main/java/ru/skypro/homework/dto/ExtendedAdDto.java
@@ -11,7 +11,7 @@ import lombok.*;
 @Builder
 public class ExtendedAdDto {
     @Schema(description = "Id объявления", example = "13221")
-    private long pk;
+    private Integer pk;
     @Schema(description = "Имя автора", example = "Иван")
     private String authorFirstName;
     @Schema(description = "Фамилия пользователя", example = "Иванов")
@@ -25,7 +25,7 @@ public class ExtendedAdDto {
     @Schema(description = "телефон автора объявления", example = "88888888888")
     private String phone;
     @Schema(description = "Цена", example = "38292,00")
-    private int price;
+    private Integer price;
     @Schema(description = "Заголовок объявления", example = "Товар")
     private String title;
 }
